refactor: split file_info structs into separate files per project convention
Move each MS-FSCC file information struct out of mod.rs into its own
file (access.rs, alignment.rs, basic.rs, ea.rs, internal.rs, mode.rs,
name.rs, network_open.rs, position.rs, standard.rs). mod.rs now only
contains module declarations, re-exports, and the composite
FileAllInformation type.

All 80 unit tests and 10 integration tests pass.
